// src/type.ts
export interface GeneralInfoData {
  name: string;
  career: string;
  summary: string;
  phone: string;
  email: string;
  address: string;
}

export interface EducationData {
  degree: string;
  year: string;
  university: string;
  major: string;
  score: string;
}

export interface ExperienceData {
  work: string;
  time: string;
  desc: string;
  desc1: string;
  desc2: string;
  work1: string;
  time1: string;
  desc3: string;
  desc4: string;
  desc5: string;
  work2: string;
  time2: string;
  desc6: string;
  desc7: string;
  desc8: string;
}